ContainsCompiler
